Provide kernel-limited.efi for provisioning
The provisioning ISO must not be signed with the production key,
as doing so makes that ISO, if you can stop the provisioning script,
usable to bypass pcr7 extension.

Note - I'm thinking of going a different route, actually.  Just
pushing this to github for safe-keeping in case the other way turns
out not to work after all.

Signed-off-by: Serge Hallyn <shallyn@cisco.com>